refactor: reorganize project structure - move modules into mangaba package
Move protocols, utils, config, and highlevel API into the mangaba package
for better organization and proper packaging. Maintains backward compatibility
through re-exports.

Changes:
- Move protocols/ to mangaba/protocols/
- Move utils/ to mangaba/utils/
- Move config.py to mangaba/config.py
- Move mangaba_agent.py to mangaba/api.py (renamed from highlevel.py)
- Create mangaba_agent.py in root for backward compatibility (re-export)
- Update all imports across the codebase
- Update pyproject.toml to only include mangaba package

Backward compatibility:
- from mangaba_agent import MangabaAgent (still works)
- from mangaba.api import MangabaAgent (new, recommended)
