+++ mangaba/utils/__init__.py
@@ -1,5 +1,5 @@
 """Utilities for Mangaba AI"""
 
-from utils.logger import get_logger
+from mangaba.utils.logger import get_logger
 
 __all__ = ["get_logger"]
