+++ pyproject.toml
@@ -123,7 +123,7 @@ Source = "https://github.com/mangaba-ai/mangaba-ai"
 Documentation = "https://github.com/mangaba-ai/mangaba-ai#readme"
 
 [tool.hatch.build.targets.wheel]
-packages = ["mangaba", "protocols", "utils"]
+packages = ["mangaba"]
 
 [tool.uv]
 # Configurações específicas para o gerenciador uv
